third_party/blink/web_tests/fast/dom/shadow/shadow-dom-event-dispatching-text-node-in-shadow-root-expected.txt



Flat Tree will be:
DIV	 id=sandbox
	DIV	 id=top
		DIV	 id=shadow-host
			#text	


Moving mouse from a direct child text node of the shadow root to top

  mouseout
     @shadow-host (target: shadow-host) (related: top)
     @top (target: shadow-host) (related: top)

  mouseover
     @top (target: top) (related: shadow-host)

  touchstart
     @shadow-host (target: shadow-host) (touches: shadow-host) (targetTouches: shadow-host) (changedTouches: shadow-host)
     @top (target: shadow-host) (touches: shadow-host) (targetTouches: shadow-host) (changedTouches: shadow-host)

  mousewheel
     @shadow-host (target: shadow-host)
     @top (target: shadow-host)
PASS successfullyParsed is true

TEST COMPLETE
